Create SQLTask concept and rewrite Retention with it
**What**
- Create new generic SQLTask spec and handler
- Rewrite the Retention implementation to use this
