+++ pkg/queue/workers/sql_task_test.go
@@ -1,0 +1,1 @@
+package workers
